pyproject.toml
[project]
name = "teampling-main-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "sqlmodel (>=0.0.37,<0.0.38)",
    "pydantic (>=2.12.5,<3.0.0)",
    "sqlalchemy-utc (>=0.14.0,<0.15.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "greenlet (>=3.3.2,<4.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
[dependency-groups]
dev = [
    "httpx (>=0.28.1,<0.29.0)"
]
